Upgrade Scala, SBT, Akka, fix errors and warnings

+++ src/main/scala/se/lu/nateko/cp/stiltweb/csv/RowCache.scala
@@ -75,7 +75,7 @@ class RowCache(rowFactory: () => Iterator[CachedRow], parentFolder: Path, year: 
 				.iterator
 				.map{minute =>
 					buff.rewind()
-					fc.read(buff, (minute / slotStepInMinutes) * BytesPerRow)
+					fc.read(buff, (minute / slotStepInMinutes) * BytesPerRow.toLong)
 					yearStart.plusMinutes(minute.toLong) -> {
 						try{
 							new String(buff.array()).parseJson.convertTo[CsvRow]
